@@ -1,34 +1,24 @@
 {
   "name": "crestron-ui",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Crestron CH5",
-  "productName": "DLCrestron",
+  "productName": "MVTCrestron",
   "author": "Dane Lewis <me@danelewis.com>",
-  "private": true,
+  "private": false,
   "scripts": {
     "lint": "eslint --ext .js,.vue ./",
     "test": "echo \"No test specified\" && exit 0",
     "build": "quasar build;ch5-cli archive -p UI-archive -d dist/spa -o archive",
-    "send": "ch5-cli deploy archive/UI-archive.ch5z -t touchscreen -H"
+    "send": "ch5-cli deploy archive/UI-archive.ch5z -t touchscreen -p -H"
   },
   "dependencies": {
-    "@crestron/ch5-crcomlib": "^1.0.0",
-    "@quasar/extras": "^1.0.0",
-    "axios": "^0.18.1",
+    "@crestron/ch5-crcomlib": "^2.10.0",
+    "@quasar/extras": "^1.16.11",
+    "axios": "^1.7.2",
     "core-js": "^3.6.5",
-    "quasar": "^1.0.0"
-  },
-  "devDependencies": {
-    "@quasar/app": "^2.0.0",
-    "babel-eslint": "^10.0.1",
-    "eslint": "^6.8.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-loader": "^3.0.3",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^6.1.2"
+    "quasar": "^2.16.4",
+    "vue": "^3.4.27",
+    "vue-router": "^4.3.2"
   },
   "browserslist": [
     "last 10 Chrome versions"
@@ -37,5 +27,17 @@
     "node": ">= 10.18.1",
     "npm": ">= 6.13.4",
     "yarn": ">= 1.21.1"
+  },
+  "devDependencies": {
+    "@babel/eslint-parser": "^7.13.14",
+    "@quasar/app-webpack": "^3.13.2",
+    "eslint": "^8.11.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.19.1",
+    "eslint-plugin-n": "^15.0.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-quasar": "^1.0.0",
+    "eslint-plugin-vue": "^9.0.0",
+    "eslint-webpack-plugin": "^3.1.1"
   }
 }
