@@ -4,20 +4,18 @@
  *
  * You are probably looking on adding startup/initialization code.
  * Use "quasar new boot <name>" and add it there.
- * One boot file per concern. Then reference the file(s) in quasar.conf.js > boot:
+ * One boot file per concern. Then reference the file(s) in quasar.config.js > boot:
  * boot: ['file', ...] // do not add ".js" extension to it.
  *
  * Boot files are your "main.js"
  **/
-import Vue from 'vue'
-import './import-quasar.js'
 
 
 
-import App from 'app/src/App.vue'
+import { Quasar } from 'quasar'
+import { markRaw } from 'vue'
+import RootComponent from 'app/src/App.vue'
 
-
-import createStore from 'app/src/store/index'
 
 import createRouter from 'app/src/router/index'
 
@@ -25,41 +23,35 @@ import createRouter from 'app/src/router/index'
 
 
 
-export default async function () {
-  // create store and router instances
-  
-  const store = typeof createStore === 'function'
-    ? await createStore({Vue})
-    : createStore
-  
-  const router = typeof createRouter === 'function'
-    ? await createRouter({Vue, store})
-    : createRouter
-  
-  // make router instance available in store
-  store.$router = router
-  
-
-  // Create the app instantiation Object.
-  // Here we inject the router, store to all child components,
-  // making them available everywhere as `this.$router` and `this.$store`.
-  const app = {
-    router,
-    store,
-    render: h => h(App)
-  }
-
+export default async function (createAppFn, quasarUserOptions) {
+  // Create the app instance.
+  // Here we inject into it the Quasar UI, the router & possibly the store.
+  const app = createAppFn(RootComponent)
 
   
-  app.el = '#q-app'
+  app.config.performance = true
   
 
-  // expose the app, the router and the store.
-  // note we are not mounting the app here, since bootstrapping will be
+  app.use(Quasar, quasarUserOptions)
+
+  
+
+  
+
+  const router = markRaw(
+    typeof createRouter === 'function'
+      ? await createRouter({})
+      : createRouter
+  )
+
+  
+
+  // Expose the app, the router and the store.
+  // Note that we are not mounting the app here, since bootstrapping will be
   // different depending on whether we are in a browser or on the server.
   return {
     app,
-    store,
+    
     router
   }
 }
