@@ -4,7 +4,7 @@
  *
  * You are probably looking on adding startup/initialization code.
  * Use "quasar new boot <name>" and add it there.
- * One boot file per concern. Then reference the file(s) in quasar.conf.js > boot:
+ * One boot file per concern. Then reference the file(s) in quasar.config.js > boot:
  * boot: ['file', ...] // do not add ".js" extension to it.
  *
  * Boot files are your "main.js"
@@ -12,32 +12,40 @@
 
 
 
-
-
 import App from 'app/src/App.vue'
-let appOptions = App.options /* Vue.extend() */ || App
-let appPrefetch = typeof appOptions.preFetch === 'function'
+let appPrefetch = typeof App.preFetch === 'function'
+  ? App.preFetch
+  : (
+    // Class components return the component options (and the preFetch hook) inside __c property
+    App.__c !== void 0 && typeof App.__c.preFetch === 'function'
+      ? App.__c.preFetch
+      : false
+    )
 
 
 function getMatchedComponents (to, router) {
   const route = to
     ? (to.matched ? to : router.resolve(to).route)
-    : router.currentRoute
+    : router.currentRoute.value
 
   if (!route) { return [] }
 
-  return Array.prototype.concat.apply([], route.matched.map(m => {
+  const matched = route.matched.filter(m => m.components !== void 0)
+
+  if (matched.length === 0) { return [] }
+
+  return Array.prototype.concat.apply([], matched.map(m => {
     return Object.keys(m.components).map(key => {
       const comp = m.components[key]
       return {
         path: m.path,
-        c: comp.options /* Vue.extend() */ || comp
+        c: comp
       }
     })
   }))
 }
 
-export function addPreFetchHooks (router, store, publicPath) {
+export function addPreFetchHooks ({ router, publicPath }) {
   // Add router hook for handling preFetch.
   // Doing it after initial route is resolved so that we don't double-fetch
   // the data that we already have. Using router.beforeResolve() so that all
@@ -57,13 +65,17 @@ export function addPreFetchHooks (router, store, publicPath) {
           m.path.indexOf('/:') > -1 // does it has params?
         ))
       })
-      .filter(m => m.c && typeof m.c.preFetch === 'function')
-      .map(m => m.c.preFetch)
+      .filter(m => m.c !== void 0 && (
+        typeof m.c.preFetch === 'function'
+        // Class components return the component options (and the preFetch hook) inside __c property
+        || (m.c.__c !== void 0 && typeof m.c.__c.preFetch === 'function')
+      ))
+      .map(m => m.c.__c !== void 0 ? m.c.__c.preFetch : m.c.preFetch)
 
     
-    if (appPrefetch === true) {
+    if (appPrefetch !== false) {
+      preFetchList.unshift(appPrefetch)
       appPrefetch = false
-      preFetchList.unshift(appOptions.preFetch)
     }
     
 
@@ -85,7 +97,7 @@ export function addPreFetchHooks (router, store, publicPath) {
 
     preFetchList.reduce(
       (promise, preFetch) => promise.then(() => hasRedirected === false && preFetch({
-        store,
+        
         currentRoute: to,
         previousRoute: from,
         redirect,
