@@ -1,4 +1,14 @@
-import Vue from 'vue'
+import { boot } from 'quasar/wrappers'
 import axios from 'axios'
 
-Vue.prototype.$axios = axios
+const api = axios.create({ baseURL: 'https://api.example.com' })
+
+export default boot(({ app }) => {
+  app.config.globalProperties.$axios = axios
+  app.config.globalProperties.$api = api
+
+  app.provide('axios', axios)
+  app.provide('api', api)
+})
+
+export { api }
