@@ -1,16 +1,18 @@
-import Vue from 'vue'
-import * as CrComLib from '@crestron/ch5-crcomlib/build_bundles/cjs/cr-com-lib'
+import { boot } from 'quasar/wrappers'
+import { CrComLib } from '@crestron/ch5-crcomlib/build_bundles/cjs/cr-com-lib'
 
-// register required components to window early
+// Register required components to window early
 window.bridgeReceiveIntegerFromNative = CrComLib.bridgeReceiveIntegerFromNative
 window.bridgeReceiveBooleanFromNative = CrComLib.bridgeReceiveBooleanFromNative
 window.bridgeReceiveStringFromNative = CrComLib.bridgeReceiveStringFromNative
 window.bridgeReceiveObjectFromNative = CrComLib.bridgeReceiveObjectFromNative
 
-// register the crestron plugin to  the vue object
 const CrestronPlugin = {
-  install: function (Vue) {
-    Object.defineProperty(Vue.prototype, '$crestron', { value: CrComLib })
+  install (app) {
+    app.config.globalProperties.$crestron = CrComLib
   }
 }
-Vue.use(CrestronPlugin)
+
+export default boot(({ app }) => {
+  app.use(CrestronPlugin)
+})
