@@ -1,4 +1,3 @@
-
 const routes = [
   {
     path: '/',
@@ -13,11 +12,8 @@ const routes = [
       { path: 'power', component: () => import('pages/Power.vue') }
     ]
   },
-
-  // Always leave this as last one,
-  // but you can also remove it
   {
-    path: '*',
+    path: '/:pathMatch(.*)*',
     component: () => import('pages/Error404.vue')
   }
 ]
